Set up ldoc to generate external docs
The intention is to have the external documentation be managed inside
the script itself, extracting and formatting it as HTML with ldoc.
Doing this means that it is easier to keep the documentation up to date
with the code, and if the code is ever distributed without the rest of
the repo, the documentation is still available.

The ldoc config sets up `macro` as an alias to `script` as "macro"
aligns more with Aegisub terminology. The template used to generate
the docs still uses the term "script", so the stock template has been
modified to use the correct term.

The template has also been modified to link to the GitHub repo
and to have the HTML title reflect the top-level page heading.

The pale theme is used instead of the default theme because it looks
better. This opinion is further supported by the fact that
even the docs for ldoc itself uses this theme.