Aegs: Prevent global styles from being deleted
The subs object that's passed to the main macro function contains the
entire ASS file, not just the dialogue events. Removing all the lines
from 1 until the aegs:end marker removed everything from the start of
the file until the marker, which includes the style definition lines.

To fix, find the first dialogue line first and use that as the start
index instead of 1 (why is there no standard util for this?).